added offers to cart summary

+++ src/test/java/com/atul/orders/model/ProductTest.java
@@ -79,7 +79,7 @@ public class ProductTest {
         field.setAccessible(true);
         field.set(product, 100);
 
-        final int result = product.getQuantity();
+        final float result = product.getQuantity();
 
         assertThat(result).isEqualTo(100);
     }
